"The data is not aligned, therefore width == stride."
Also, moved all the svg code into svg.rs and got it out of the mod file.
